move demo function to the component
